requirements.txt
streamlit>=1.28
pandas
requests
docxtpl
openpyxl
pypdf
num2words
streamlit-searchbox
python-dateutil
supabase
python-dotenv
